backend/requirements.txt
django
djangorestframework
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt
social-auth-app-django
google-auth
google-auth-oauthlib
google-auth-httplib2
Pillow
channels
channels-redis